# Backend/Frontend/openapi.raw.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, an unknown field was sent, or role is TECHNICIAN (use POST /technicians instead)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "409": {
            "description": "username or email already in use"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a user account (ADMIN only)",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by User.isActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial match on username, email, firstName or lastName",
            "schema": {
              "minLength": 1,
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List user accounts, paginated and filtered, sorted by username ASC (ADMIN only). Soft-deleted accounts are excluded.",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a single user account by id (ADMIN only)",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No field provided, a role change to/from TECHNICIAN, or an attempt by the caller to change their own role/activation"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "username or email already in use"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a user account — profile fields, role (ADMIN \u003C-\u003E CLIENT only) and activation (ADMIN only)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Account soft-deleted"
          },
          "400": {
            "description": "An ADMIN cannot delete themselves"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "The user is still assigned to open tickets"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a user account (ADMIN only). The row is kept so tickets and history stay traceable.",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the currently authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ticket Checker API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "TECHNICIAN",
          "CLIENT"
        ],
        "description": "ADMIN or CLIENT. TECHNICIAN is rejected with 400 — use POST /technicians, which creates the account and its technician profile in one transaction (D1)."
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "jdoe",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "example": "jdoe@example.com"
          },
          "password": {
            "type": "string",
            "example": "Str0ngP@ssw0rd",
            "minLength": 10,
            "maxLength": 72
          },
          "role": {
            "default": "CLIENT",
            "description": "ADMIN or CLIENT. TECHNICIAN is rejected with 400 — use POST /technicians, which creates the account and its technician profile in one transaction (D1).",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "firstName": {
            "type": "string",
            "example": "Jane",
            "maxLength": 80
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "example": "+228 90 00 00 00",
            "maxLength": 30
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "firstName",
          "lastName",
          "phone",
          "role",
          "isActive",
          "createdAt"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total number of items across all pages"
          },
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current, 1-based page number"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "example": 3,
            "description": "Total number of pages"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Page of results. The concrete item type is documented per-endpoint by @ApiPaginatedResponse()."
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "maxLength": 80
          },
          "lastName": {
            "type": "string",
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "maxLength": 30
          },
          "role": {
            "description": "ADMIN \u003C-\u003E CLIENT only. Any change to or from TECHNICIAN is rejected with 400 (D2).",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean",
            "description": "Deactivating an account takes effect on the very next request: JwtStrategy re-reads the user and rejects an inactive one (D5)."
          }
        }
      },
...
    }
  }
}